Improved documentation and commands for maintainability
