+++ .github/prompts/review-pr.md
@@ -1,0 +1,47 @@
+A variety of comments have been raised on this PR — please take a look.
+
+https://github.com/pedrazamiguez/split-trip/pull/xxxx
+
+---
+
+## Step 1 — Read and triage all comments
+
+For each review comment or thread:
+
+1. **Understand the context** — read the surrounding code, not just the highlighted line.
+2. **Decide the outcome:**
+   - ✅ **Valid** — address it with a code change.
+   - ℹ️ **Partially valid** — address the spirit of the comment, explain any deviation in a reply.
+   - ❌ **False positive** — reply with a clear technical reason why no change is needed.
+3. **Reply to every comment**, even if no code change is made. Never leave a thread unanswered.
+
+---
+
+## Step 2 — File-size guard before editing
+
+Before touching any file, check its current line count:
+
+```bash
+wc -l <path/to/file.kt>
+```
+
+If a file is at or near 600 lines (the Konsist hard limit), plan a split or extraction **before** adding code. After editing, re-check and refactor immediately if over 600 lines.
+
+---
+
+## Step 3 — Implement changes
+
+Follow all architecture constraints in `.github/copilot-instructions.md` and `AGENTS.md`.  
+Take into account detekt rules, ktlint formatting, and the 600-line file-size limit.
+
+---
+
+## Step 4 — Local verification gate (run BEFORE declaring done)
+
+Do not consider the review addressed until ALL of the following pass locally:
+
+```bash
+make check   # Konsist architecture rules + all unit tests + debug compilation — must show 0 failures
+```
+
+If any check fails, fix it before finishing. Do not leave failures for CI to catch.
