+++ Makefile
@@ -1,0 +1,140 @@
+# ─────────────────────────────────────────────────────────────────────────────
+# SplitTrip — Developer Bootstrap & Quality Gates
+#
+# Run once after cloning or switching machines:
+#   make setup
+#
+# Run before pushing:
+#   make check
+# ─────────────────────────────────────────────────────────────────────────────
+
+GRADLEW       := ./gradlew
+LOCAL_PROPS   := local.properties
+HOOK_SRC      := scripts/pre-commit
+HOOK_DST      := .git/hooks/pre-commit
+
+# ─── Colours ─────────────────────────────────────────────────────────────────
+RED    := \033[0;31m
+GREEN  := \033[0;32m
+YELLOW := \033[1;33m
+CYAN   := \033[0;36m
+NC     := \033[0m
+
+.PHONY: help setup hooks local-props doctor check test konsist build clean
+
+# ─── Default: show help ───────────────────────────────────────────────────────
+help: ## Show this help message
+	@echo ""
+	@printf "$(CYAN)SplitTrip — Developer Bootstrap$(NC)\n"
+	@echo ""
+	@printf "$(YELLOW)Bootstrap (run once after cloning / switching machines):$(NC)\n"
+	@printf "  $(CYAN)make setup$(NC)      Full first-time setup\n"
+	@echo ""
+	@printf "$(YELLOW)Available targets:$(NC)\n"
+	@grep -E '^[a-zA-Z_-]+:.*?## .*$$' $(MAKEFILE_LIST) | \
+		awk 'BEGIN {FS = ":.*?## "}; {printf "  $(CYAN)%-16s$(NC) %s\n", $$1, $$2}'
+	@echo ""
+
+# ─── Bootstrap ────────────────────────────────────────────────────────────────
+setup: hooks local-props doctor ## Full first-time setup: install hooks, scaffold local.properties, run doctor
+	@echo ""
+	@printf "$(GREEN)✅  Setup complete.$(NC)\n"
+	@echo ""
+	@printf "$(YELLOW)Next steps (if not done already):$(NC)\n"
+	@printf "  1. Copy google-services.json into app/\n"
+	@printf "  2. Add OER_APP_ID_DEBUG and OER_APP_ID_RELEASE to ~/.gradle/gradle.properties\n"
+	@printf "  3. Open the project in Android Studio and sync Gradle\n"
+	@echo ""
+
+hooks: ## Install the pre-commit git hook (no Gradle needed)
+	@if [ ! -d ".git" ]; then \
+		printf "$(RED)✗  .git directory not found — are you inside the repo root?$(NC)\n"; \
+		exit 1; \
+	fi
+	@mkdir -p .git/hooks
+	@cp $(HOOK_SRC) $(HOOK_DST)
+	@chmod +x $(HOOK_DST)
+	@printf "$(GREEN)✅  Pre-commit hook installed → $(HOOK_DST)$(NC)\n"
+
+local-props: ## Create local.properties (sdk.dir only) if it does not exist
+	@if [ ! -f "$(LOCAL_PROPS)" ]; then \
+		printf "$(YELLOW)⏳  Creating local.properties...$(NC)\n"; \
+		echo "# Auto-generated by 'make setup'" > $(LOCAL_PROPS); \
+		echo "# This file is for sdk.dir ONLY — API keys do NOT go here." >> $(LOCAL_PROPS); \
+		echo "# See: ~/.gradle/gradle.properties for OER_APP_ID_DEBUG / OER_APP_ID_RELEASE" >> $(LOCAL_PROPS); \
+		if [ -n "$$ANDROID_HOME" ]; then \
+			echo "sdk.dir=$$ANDROID_HOME" >> $(LOCAL_PROPS); \
+			printf "$(GREEN)✅  local.properties created$(NC) (sdk.dir set from \$$ANDROID_HOME)\n"; \
+		else \
+			echo "sdk.dir=/path/to/your/android/sdk" >> $(LOCAL_PROPS); \
+			printf "$(GREEN)✅  local.properties created$(NC) — update sdk.dir (or set \$$ANDROID_HOME and re-run)\n"; \
+		fi; \
+	else \
+		printf "$(GREEN)✅  local.properties already exists — skipping$(NC)\n"; \
+	fi
+
+doctor: ## Check that required files and tools are present
+	@echo ""
+	@printf "$(YELLOW)🩺  Running doctor checks...$(NC)\n"
+	@echo ""
+	@# JDK 21
+	@if java -version 2>&1 | grep -q 'version "21'; then \
+		printf "  $(GREEN)✅  JDK 21 detected$(NC)\n"; \
+	else \
+		printf "  $(RED)✗   JDK 21 not found$(NC) — install via SDKMAN or your package manager\n"; \
+	fi
+	@# google-services.json
+	@if [ -f "app/google-services.json" ]; then \
+		printf "  $(GREEN)✅  app/google-services.json present$(NC)\n"; \
+	else \
+		printf "  $(YELLOW)⚠️   app/google-services.json missing$(NC) — download from Firebase Console\n"; \
+	fi
+	@# local.properties — sdk.dir
+	@if grep -q "^sdk.dir=" $(LOCAL_PROPS) 2>/dev/null; then \
+		SDK_PATH=$$(grep "^sdk.dir=" $(LOCAL_PROPS) | cut -d= -f2); \
+		if [ -d "$$SDK_PATH" ]; then \
+			printf "  $(GREEN)✅  Android SDK found at $$SDK_PATH$(NC)\n"; \
+		else \
+			printf "  $(RED)✗   sdk.dir path does not exist: $$SDK_PATH$(NC)\n"; \
+		fi; \
+	else \
+		printf "  $(RED)✗   sdk.dir not set in local.properties$(NC)\n"; \
+	fi
+	@# OER API key — must be in ~/.gradle/gradle.properties (not local.properties)
+	@if grep -q "^OER_APP_ID_DEBUG=" ~/.gradle/gradle.properties 2>/dev/null && \
+		! grep -q "your_" ~/.gradle/gradle.properties 2>/dev/null; then \
+		printf "  $(GREEN)✅  OER_APP_ID_DEBUG is set in ~/.gradle/gradle.properties$(NC)\n"; \
+	else \
+		printf "  $(YELLOW)⚠️   OER_APP_ID_DEBUG not set$(NC) — add to ~/.gradle/gradle.properties:\n"; \
+		printf "         OER_APP_ID_DEBUG=your_key_here\n"; \
+		printf "         OER_APP_ID_RELEASE=your_key_here\n"; \
+		printf "       $(YELLOW)(not local.properties — those are Gradle properties, not Android properties)$(NC)\n"; \
+	fi
+	@# pre-commit hook
+	@if [ -x "$(HOOK_DST)" ]; then \
+		printf "  $(GREEN)✅  Pre-commit hook installed$(NC)\n"; \
+	else \
+		printf "  $(YELLOW)⚠️   Pre-commit hook not installed$(NC) — run 'make hooks'\n"; \
+	fi
+	@echo ""
+
+# ─── Quality gates (mirrors CI) ───────────────────────────────────────────────
+check: konsist test build ## Run all local quality gates before pushing (mirrors CI)
+
+test: ## Run all unit tests
+	@printf "$(YELLOW)⏳  Running unit tests...$(NC)\n"
+	@$(GRADLEW) test --continue
+
+konsist: ## Run Konsist architecture tests (file-size limit, naming, dependency rules)
+	@printf "$(YELLOW)⏳  Running Konsist architecture tests...$(NC)\n"
+	@$(GRADLEW) :konsist-tests:test
+
+build: ## Compile all modules (debug)
+	@printf "$(YELLOW)⏳  Compiling debug...$(NC)\n"
+	@$(GRADLEW) compileDebugKotlin --continue
+
+# ─── Utility ──────────────────────────────────────────────────────────────────
+clean: ## Clean all Gradle build outputs
+	@printf "$(YELLOW)⏳  Cleaning build outputs...$(NC)\n"
+	@$(GRADLEW) clean
+
