+++ features/expenses/src/main/kotlin/es/pedrazamiguez/splittrip/features/expense/presentation/viewmodel/handler/CurrencyEventHandler.kt
@@ -3,20 +3,15 @@ package es.pedrazamiguez.splittrip.features.expense.presentation.viewmodel.handl
 import es.pedrazamiguez.splittrip.core.common.presentation.UiText
 import es.pedrazamiguez.splittrip.core.designsystem.presentation.formatter.FormattingHelper
 import es.pedrazamiguez.splittrip.domain.enums.PayerType
-import es.pedrazamiguez.splittrip.domain.model.CashRatePreviewResult
+import es.pedrazamiguez.splittrip.domain.enums.PaymentMethod
 import es.pedrazamiguez.splittrip.domain.service.ExchangeRateCalculationService
-import es.pedrazamiguez.splittrip.domain.service.ExpenseCalculatorService
-import es.pedrazamiguez.splittrip.domain.service.split.SplitPreviewService
 import es.pedrazamiguez.splittrip.domain.usecase.currency.GetExchangeRateUseCase
-import es.pedrazamiguez.splittrip.domain.usecase.expense.PreviewCashExchangeRateUseCase
 import es.pedrazamiguez.splittrip.features.expense.R
 import es.pedrazamiguez.splittrip.features.expense.presentation.mapper.AddExpenseOptionsUiMapper
 import es.pedrazamiguez.splittrip.features.expense.presentation.viewmodel.action.AddExpenseUiAction
 import es.pedrazamiguez.splittrip.features.expense.presentation.viewmodel.state.AddExpenseUiState
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Job
-import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.update
@@ -27,43 +22,27 @@ import timber.log.Timber
  * Handles currency selection and exchange rate events:
  * [CurrencySelected], [ExchangeRateChanged], [GroupAmountChanged].
  *
- * Also exposes [fetchRate] and [recalculateForward] for cross-handler calls
- * (e.g., from [ConfigEventHandler] after initial config load).
+ * Also exposes [fetchRate], [fetchCashRate], [recalculateForward], and [recalculateCashForward]
+ * for cross-handler calls (e.g., from [ConfigEventHandler] after initial config load).
+ *
+ * CASH-specific async operations (rate fetching, result mapping, debounce) are delegated
+ * to [CashRateDelegate] to stay within the 600-line production-file limit.
  */
 // Function count driven by event/action categories (CASH, non-CASH, funding source);
-// extracting further would require a Delegate sub-pattern
+// extracting further would require additional Delegate sub-patterns
 @Suppress("TooManyFunctions")
 class CurrencyEventHandler(
     private val getExchangeRateUseCase: GetExchangeRateUseCase,
-    private val previewCashExchangeRateUseCase: PreviewCashExchangeRateUseCase,
     private val exchangeRateCalculationService: ExchangeRateCalculationService,
-    private val expenseCalculatorService: ExpenseCalculatorService,
-    private val splitPreviewService: SplitPreviewService,
     private val formattingHelper: FormattingHelper,
     private val addExpenseOptionsMapper: AddExpenseOptionsUiMapper,
-    private val withdrawalPoolSelectionDelegate: WithdrawalPoolSelectionDelegate
+    private val withdrawalPoolSelectionDelegate: WithdrawalPoolSelectionDelegate,
+    private val cashRateDelegate: CashRateDelegate
 ) : AddExpenseEventHandler {
 
     private lateinit var _uiState: MutableStateFlow<AddExpenseUiState>
     private lateinit var _actions: MutableSharedFlow<AddExpenseUiAction>
     private lateinit var scope: CoroutineScope
-
-    /** Debounce job for cash rate preview recalculations on amount changes. */
-    private var cashPreviewJob: Job? = null
-
-    /** Tracked coroutine for in-flight cash rate fetch (prevents stale/duplicate results). */
-    private var cashRateJob: Job? = null
-
-    companion object {
-        private const val CASH_PREVIEW_DEBOUNCE_MS = 300L
-
-        /**
-         * Placeholder shown in locked exchange-rate fields when no value is available
-         * (e.g. insufficient cash, no withdrawals). Keeps the OutlinedTextField label
-         * floating above the field instead of collapsing into the field body.
-         */
-        private const val EMPTY_FIELD_PLACEHOLDER = "—"
-    }
 
     override fun bind(
         stateFlow: MutableStateFlow<AddExpenseUiState>,
@@ -73,6 +52,7 @@ class CurrencyEventHandler(
         _uiState = stateFlow
         _actions = actionsFlow
         this.scope = scope
+        cashRateDelegate.setup(stateFlow, scope)
     }
 
     fun handleCurrencySelected(currencyCode: String, onRecalculate: () -> Unit) {
@@ -111,7 +91,7 @@ class CurrencyEventHandler(
                         )
                     )
                 }
-                fetchCashRate()
+                cashRateDelegate.fetchCashRate()
             } else {
                 fetchRate()
             }
@@ -125,7 +105,7 @@ class CurrencyEventHandler(
                 )
             }
             // Same currency + CASH + GROUP pocket: fetch tranche preview (shown on AmountStep)
-            if (isCash && currentPayerType() == PayerType.GROUP) fetchCashRate()
+            if (isCash && currentPayerType() == PayerType.GROUP) cashRateDelegate.fetchCashRate()
         }
         recalculateForward()
         onRecalculate()
@@ -268,7 +248,7 @@ class CurrencyEventHandler(
                     )
                 }
                 withdrawalPoolSelectionDelegate.clearPoolState(_uiState)
-                fetchCashRate()
+                cashRateDelegate.fetchCashRate()
             } else {
                 // USER/SUBUNIT cash: rate stays unlocked, user enters manually
                 // but we still fetch pool availability for pool-selector widget.
@@ -277,11 +257,11 @@ class CurrencyEventHandler(
         } else if (isCash && !isForeign && isGroupPocket) {
             // Same-currency CASH: no rate to lock, but still fetch tranche preview
             withdrawalPoolSelectionDelegate.clearPoolState(_uiState)
-            fetchCashRate()
+            cashRateDelegate.fetchCashRate()
         } else {
             // Cancel any in-flight or debounced CASH rate jobs so a stale result
             // cannot re-lock the exchange rate after the user has switched away.
-            cancelPendingCashJobs()
+            cashRateDelegate.cancelPendingCashJobs()
             withdrawalPoolSelectionDelegate.clearPoolState(_uiState)
 
             _uiState.update {
@@ -345,10 +325,10 @@ class CurrencyEventHandler(
                     }
                 }
                 withdrawalPoolSelectionDelegate.clearPoolState(_uiState)
-                fetchCashRate()
+                cashRateDelegate.fetchCashRate()
             } else {
                 // Switching to USER/SUBUNIT: cancel ATM jobs, unlock rate, restore saved rate
-                cancelPendingCashJobs()
+                cashRateDelegate.cancelPendingCashJobs()
                 val savedRate = state.preCashExchangeRate
                 _uiState.update {
                     it.copy(
@@ -371,10 +351,10 @@ class CurrencyEventHandler(
             if (isGroupPocket) {
                 // Switching to GROUP pocket: (re)fetch tranche preview for the current amount.
                 withdrawalPoolSelectionDelegate.clearPoolState(_uiState)
-                fetchCashRate()
+                cashRateDelegate.fetchCashRate()
             } else {
                 // Switching to USER/SUBUNIT pocket: clear tranche state — ATM pool is not used.
-                cancelPendingCashJobs()
+                cashRateDelegate.cancelPendingCashJobs()
                 withdrawalPoolSelectionDelegate.clearPoolState(_uiState)
                 _uiState.update {
                     it.copy(
@@ -386,209 +366,11 @@ class CurrencyEventHandler(
         }
     }
 
-    /**
-     * Fetches the blended ATM exchange rate preview for the current source currency and amount.
-     * Shows a weighted-average when no amount is entered yet.
-     * Cancels any previous in-flight request and applies a stale-result guard.
-     *
-     * When the user has explicitly selected a pool via the pool-selection widget,
-     * delegates to that pool's exact scope by passing [preferredWithdrawalScope] and
-     * [preferredWithdrawalOwnerId] to [PreviewCashExchangeRateUseCase]. Otherwise,
-     * uses the automatic scope-priority logic (personal pool first, GROUP fallback).
-     */
-    fun fetchCashRate() {
-        val state = _uiState.value
-        val groupId = state.loadedGroupId ?: return
-        val sourceCurrency = state.selectedCurrency?.code ?: return
-        val groupCurrency = state.groupCurrency?.code ?: return
+    /** Thin wrapper — delegates to [CashRateDelegate.fetchCashRate]. */
+    fun fetchCashRate() = cashRateDelegate.fetchCashRate()
 
-        val isSameCurrency = sourceCurrency == groupCurrency
-        val sourceDecimalDigits = state.selectedCurrency.decimalDigits
-        val targetDecimalDigits = state.groupCurrency.decimalDigits
-
-        // Parse current source amount to cents (0 if blank/invalid)
-        val sourceAmountCents = splitPreviewService.parseAmountToCents(
-            state.sourceAmount,
-            sourceDecimalDigits
-        )
-
-        // Resolve scope context from the current funding source selection
-        val payerType = currentPayerType()
-        val payerId = currentPayerId()
-
-        // If the user has explicitly selected a pool, use that pool's exact scope for the preview.
-        val selectedPool = state.selectedWithdrawalPool
-        val preferredScope = selectedPool?.scope
-        val preferredOwnerId = selectedPool?.ownerId
-
-        // Capture request context for stale-result check
-        val requestedGroupId = groupId
-        val requestedSourceCurrency = sourceCurrency
-
-        // Cancel any previous in-flight cash rate request
-        cashRateJob?.cancel()
-        cashRateJob = scope.launch {
-            _uiState.update { it.copy(isLoadingRate = true) }
-            try {
-                val result = previewCashExchangeRateUseCase(
-                    groupId = requestedGroupId,
-                    sourceCurrency = requestedSourceCurrency,
-                    sourceAmountCents = sourceAmountCents,
-                    payerType = payerType,
-                    payerId = payerId,
-                    preferredWithdrawalScope = preferredScope,
-                    preferredWithdrawalOwnerId = preferredOwnerId
-                )
-
-                _uiState.update { current ->
-                    // Stale-result check: ignore if the user changed group or currency
-                    // while the request was in-flight.
-                    if (current.loadedGroupId != requestedGroupId ||
-                        current.selectedCurrency?.code != requestedSourceCurrency
-                    ) {
-                        return@update current.copy(isLoadingRate = false)
-                    }
-
-                    mapCashRateResult(current, result, targetDecimalDigits, requestedSourceCurrency, !isSameCurrency)
-                }
-            } catch (e: Exception) {
-                Timber.e(e, "Failed to preview cash exchange rate")
-                _uiState.update { it.copy(isLoadingRate = false) }
-            }
-        }
-    }
-
-    /**
-     * Maps a [CashRatePreviewResult] into an updated [AddExpenseUiState].
-     *
-     * Extracted from [fetchCashRate] to reduce method length and complexity.
-     */
-    // Sealed when-expression maps each result variant to a state copy;
-    // length is driven by the number of fields per branch, not logic
-    @Suppress("LongMethod")
-    internal fun mapCashRateResult(
-        current: AddExpenseUiState,
-        result: CashRatePreviewResult,
-        targetDecimalDigits: Int,
-        sourceCurrencyCode: String = "",
-        updateRateFields: Boolean = true
-    ): AddExpenseUiState = when (result) {
-        is CashRatePreviewResult.Available -> {
-            val preview = result.preview
-            val tranchePreviews = if (preview.tranches.isNotEmpty() && sourceCurrencyCode.isNotBlank()) {
-                addExpenseOptionsMapper.mapCashTranchePreviews(preview.tranches, sourceCurrencyCode)
-            } else {
-                persistentListOf()
-            }
-
-            // Same-currency CASH: only update tranche preview and clear insufficient flag.
-            if (!updateRateFields) {
-                return current.copy(
-                    isLoadingRate = false,
-                    isInsufficientCash = false,
-                    cashTranchePreviews = tranchePreviews
-                )
-            }
-
-            val formattedRate = formattingHelper.formatRateForDisplay(
-                preview.displayRate.toPlainString()
-            )
-
-            if (preview.groupAmountCents > 0) {
-                // FIFO-simulated: update both rate and group amount
-                val groupAmountStr = expenseCalculatorService.centsToBigDecimalString(
-                    preview.groupAmountCents,
-                    targetDecimalDigits
-                )
-                val formattedAmount = formattingHelper.formatForDisplay(
-                    internalValue = groupAmountStr,
-                    maxDecimalPlaces = targetDecimalDigits,
-                    minDecimalPlaces = targetDecimalDigits
-                )
-                current.copy(
-                    isLoadingRate = false,
-                    displayExchangeRate = formattedRate,
-                    calculatedGroupAmount = formattedAmount,
-                    isExchangeRateLocked = true,
-                    isInsufficientCash = false,
-                    exchangeRateLockedHint = UiText.StringResource(
-                        R.string.add_expense_cash_rate_locked_hint
-                    ),
-                    cashTranchePreviews = tranchePreviews
-                )
-            } else {
-                // Weighted-average preview (no amount entered yet).
-                current.copy(
-                    isLoadingRate = false,
-                    displayExchangeRate = formattedRate,
-                    calculatedGroupAmount = "",
-                    isExchangeRateLocked = true,
-                    isInsufficientCash = false,
-                    exchangeRateLockedHint = UiText.StringResource(
-                        R.string.add_expense_cash_rate_locked_hint
-                    ),
-                    cashTranchePreviews = persistentListOf()
-                )
-            }
-        }
-
-        is CashRatePreviewResult.InsufficientCash -> {
-            current.copy(
-                isLoadingRate = false,
-                displayExchangeRate = if (updateRateFields) {
-                    EMPTY_FIELD_PLACEHOLDER
-                } else {
-                    current.displayExchangeRate
-                },
-                calculatedGroupAmount = if (updateRateFields) {
-                    EMPTY_FIELD_PLACEHOLDER
-                } else {
-                    current.calculatedGroupAmount
-                },
-                isExchangeRateLocked = updateRateFields || current.isExchangeRateLocked,
-                isInsufficientCash = true,
-                exchangeRateLockedHint = if (updateRateFields) {
-                    UiText.StringResource(R.string.add_expense_cash_insufficient_hint)
-                } else {
-                    current.exchangeRateLockedHint
-                },
-                cashTranchePreviews = persistentListOf()
-            )
-        }
-
-        is CashRatePreviewResult.NoWithdrawals -> {
-            current.copy(
-                isLoadingRate = false,
-                displayExchangeRate = if (updateRateFields) {
-                    EMPTY_FIELD_PLACEHOLDER
-                } else {
-                    current.displayExchangeRate
-                },
-                calculatedGroupAmount = if (updateRateFields) {
-                    EMPTY_FIELD_PLACEHOLDER
-                } else {
-                    current.calculatedGroupAmount
-                },
-                isExchangeRateLocked = updateRateFields,
-                isInsufficientCash = false,
-                exchangeRateLockedHint = if (updateRateFields) {
-                    UiText.StringResource(R.string.add_expense_cash_rate_locked_hint)
-                } else {
-                    current.exchangeRateLockedHint
-                },
-                cashTranchePreviews = persistentListOf()
-            )
-        }
-    }
-
-    /** Debounced CASH rate recalculation — avoids hitting Room on every keystroke. */
-    fun recalculateCashForward() {
-        cashPreviewJob?.cancel()
-        cashPreviewJob = scope.launch {
-            delay(CASH_PREVIEW_DEBOUNCE_MS)
-            fetchCashRate()
-        }
-    }
+    /** Thin wrapper — delegates to [CashRateDelegate.recalculateCashForward]. */
+    fun recalculateCashForward() = cashRateDelegate.recalculateCashForward()
 
     /**
      * Handles the user's explicit pool selection from the pool-selection widget.
@@ -597,7 +379,7 @@ class CurrencyEventHandler(
      */
     fun handleWithdrawalPoolSelected(scope: PayerType, scopeOwnerId: String?) {
         withdrawalPoolSelectionDelegate.handlePoolSelected(scope, scopeOwnerId, _uiState) {
-            fetchCashRate()
+            cashRateDelegate.fetchCashRate()
         }
     }
 
@@ -619,20 +401,8 @@ class CurrencyEventHandler(
             payerId = payerId,
             scope = scope,
             stateFlow = _uiState,
-            onPoolResolved = ::fetchCashRate
+            onPoolResolved = cashRateDelegate::fetchCashRate
         )
-    }
-
-    /**
-     * Cancels any in-flight or debounced CASH rate jobs.
-     * Called when leaving CASH payment to prevent a stale result from
-     * re-locking the exchange rate after the user has switched away.
-     */
-    private fun cancelPendingCashJobs() {
-        cashRateJob?.cancel()
-        cashRateJob = null
-        cashPreviewJob?.cancel()
-        cashPreviewJob = null
     }
 
     /**
@@ -641,8 +411,7 @@ class CurrencyEventHandler(
     private fun isCashPaymentMethod(): Boolean {
         val methodId = _uiState.value.selectedPaymentMethod?.id ?: return false
         return try {
-            es.pedrazamiguez.splittrip.domain.enums.PaymentMethod.fromString(methodId) ==
-                es.pedrazamiguez.splittrip.domain.enums.PaymentMethod.CASH
+            PaymentMethod.fromString(methodId) == PaymentMethod.CASH
         } catch (_: IllegalArgumentException) {
             false
         }
